Seperate commands to cog categories
